adjusted signup thunk to include code
